refactor: функция creat_text перенесена в отдельный модуль
