+++ handlers/pictures.py
@@ -13,51 +13,15 @@ from keyboards.keyboards import (
     menu_kb,
     skip
 )
+from keyboards.functions import creat_text
 from config_data.config import config
 
 from aiogram.fsm.context import FSMContext
 from fsm.fsm import FSM_PICTURE
 from aiogram.fsm.state import default_state
 
+
 router = Router()
-
-
-# функция формирования сообшения
-def creat_text(users_db, id, mode: str, **kwargs) -> str:
-    if mode == 'ready':
-        if 'enter_telephone' in kwargs:
-            text = f'Заказ: Готовая картина\n\n' \
-                   f'Имя: {users_db[id]["name"]}\n' \
-                   f'Телефон: {kwargs["enter_telephone"]}\n' \
-                   f'Способ связи: {kwargs["how_contact"]}'
-
-        elif 'enter_email' in kwargs:
-            text = f'Заказ: Готовая картина\n\n' \
-                   f'Имя: {users_db[id]["name"]}\n' \
-                   f'E-mail: {kwargs["enter_email"]}'
-
-    elif mode == 'order':
-        if "enter_telephone" in kwargs:
-            text = f'Заказ картины\n\n' \
-                   f'Имя: {users_db[id]["name"]}\n' \
-                   f'Телефон: {kwargs["enter_telephone"]}\n' \
-                   f'Способ связи: {kwargs["how_contact"]}\n' \
-                   f'Кому картина: {kwargs.get("for_whom", "   ---")}\n' \
-                   f'Событие: {kwargs.get("event", "   ---")}\n' \
-                   f'Размер: {kwargs.get("size", "   ---")}\n' \
-                   f'Настроение: {kwargs.get("mood", "   ---")}\n' \
-                   f'Цвета: {kwargs.get("color", "   ---")}'
-
-        elif "enter_email" in kwargs:
-            text = f'Заказ картины\n\n' \
-                   f'Имя: {users_db[id]["name"]}\n' \
-                   f'Email: {kwargs["enter_email"]}\n' \
-                   f'Кому картина: {kwargs.get("for_whom", "   ---")}\n' \
-                   f'Событие: {kwargs.get("event", "   ---")}\n' \
-                   f'Размер: {kwargs.get("size", "   ---")}\n' \
-                   f'Настроение: {kwargs.get("mood", "   ---")}\n' \
-                   f'Цвета: {kwargs.get("color", "   ---")}'
-    return text
 
 
 # Хендлер на кнопку меню 'Картины'
@@ -235,7 +199,7 @@ async def mood(message: Message, state: FSMContext):
     await state.set_state(FSM_PICTURE.color)
 
 
-# Хендлер на вопрос 'Цветовая гамма' - отправка даннных админам
+# Хендлер на вопрос 'Цветовая гамма' - проверка всех введенных данных пользователем
 @router.message(StateFilter(FSM_PICTURE.color))
 @router.message(StateFilter(FSM_PICTURE.color), F.text == LEXICON_PICTURES['skip'])
 async def color(message: Message, state: FSMContext):
